% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{add_assessments}
\alias{add_assessments}
\title{Add model assessments}
\usage{
add_assessments(regression_model, evaluation_methods, is_lmer = FALSE)
}
\arguments{
\item{regression_model}{Regression model that is assessed}

\item{evaluation_methods}{List of methods to use for model evaluation Options are: anova, aic, aicc, and bic}

\item{is_lmer}{Boolean stating whether model is of type lmer}
}
\value{
Regression model with added assessments based on assessment methods provided
}
\description{
Add evaluations of model given model and evaluation methods (anova/aic/aicc/bic)
}
