% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{add_letters}
\alias{add_letters}
\title{Add letters to factor levels}
\usage{
add_letters(response, predictor, m_frame, stat_result)
}
\arguments{
\item{response}{String representing the response/y variable (column name)}

\item{predictor}{String representing the predictor/x variable (column name)}

\item{m_frame}{Dataframe with response and predictors as columns}

\item{stat_result}{Results of either wilcox or t.test}
}
\value{
Updated dataframe with letters
}
\description{
Adds letters indicating significant differences between factor levels
}
