% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{apply_psi}
\alias{apply_psi}
\title{Calculate post-selection inference}
\usage{
apply_psi(
  df,
  regression_model,
  model_type,
  model_family,
  term,
  categorical_vars,
  models_overview,
  boot_repl = 100,
  k = 2,
  round_p = 6
)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns.}

\item{regression_model}{A starter glm model with a complete term.}

\item{model_type}{The model to be used.
Options: glm or lm.}

\item{model_family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options.}

\item{term}{The call to be used with the model.}

\item{categorical_vars}{List of categorical variables within the model
(base names, i.e., names of columns)}

\item{models_overview}{Output of \code{\link[=expand_model_summary]{expand_model_summary()}}.}

\item{boot_repl}{A number or list of bootstrap replicates.
The default is no bootstrapping.}

\item{k}{The multiple of the number of degrees of freedom used as
penalty in the model selection. The default k = 2 corresponds to the AIC.}

\item{round_p}{Convenience parameter for automatic rounding of p-values.}
}
\value{
Returns a dataframe including the model's independent variables and
their respective raw and expected p-values
as well as the updated glm model
}
\description{
Calculate post-selection inference for glm and lm models
}
