% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{assess_basic_model_quality}
\alias{assess_basic_model_quality}
\title{Assess basic model quality}
\usage{
assess_basic_model_quality(model, quality_assessment, model_type)
}
\arguments{
\item{model}{A model for which to plot features}

\item{quality_assessment}{The mode of model quality assessment. Either "baseR" or "performance"}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "gam"}
}
\value{
Either a grid with plots or NA if quality_assessment differed from "baseR"
and "performance"
}
\description{
Assess basic model quality
}
