% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{backward_simplification}
\alias{backward_simplification}
\title{Backward regression model simplification}
\usage{
backward_simplification(
  df,
  model_type,
  term,
  evaluation_methods,
  categorical_vars,
  ...,
  model_family = "gaussian",
  simplify_model = TRUE,
  trace = FALSE,
  omit_na = "overall"
)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{model_type}{The model to be used.
Options: (g)lm, (g)lmer, nlmer, and gam. Default: glm.}

\item{term}{The formula to be used with the model.
Can be either quote() or formula()}

\item{evaluation_methods}{Methods to be used for model evaluation.
Options: anova, aic, aicc, bic, default: anova.}

\item{categorical_vars}{List of categorical variables within the model
(base names, i.e., names of columns)}

\item{...}{Parameters to be directly used with model call}

\item{model_family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Default: gaussian.}

\item{simplify_model}{If FALSE, the model and information on significant and
marginally significant variables are returned without any improvements.}

\item{trace}{Store and return model selection history (default: FALSE)}

\item{omit_na}{Either 'overall' or 'stepwise'.
If 'overall', NAs are removed before modeling.
If 'stepwise', NAs are removed per step based on
the variables in the current formula}
}
\value{
List containing the final regression model,
the significant and marginally significant model variables
and, if trace is TRUE, the selection history
}
\description{
Simplify the regression model by eliminating model variables one by one
}
