% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_base_formula}
\alias{check_base_formula}
\title{Check base formula used for forward model selection}
\usage{
check_base_formula(base_formula, formula)
}
\arguments{
\item{base_formula}{The lower formula used for forward model selection. Only required if
direction = "forward", otherwise this is NA}

\item{formula}{Upper formula to be used for model creation/selection.
Used for extraction of response.}
}
\value{
A base formula to be used for forward model selection
}
\description{
Checks whether base formula was provided and creates one if
necessary.
}
