% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_cbind_model_family}
\alias{check_cbind_model_family}
\title{Checks model family in case of \code{\link[=cbind]{cbind()}}}
\usage{
check_cbind_model_family(model_type, data, pasted_response)
}
\arguments{
\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "lmer", "glmer",
"nlme", "gam", and "nls".}

\item{data}{The underlying data to use when determining an appropriate model family}

\item{pasted_response}{Response as character}
}
\value{
A list containing all valid families and a notice explaining
the family choice
}
\description{
Checks which model family matches data when \code{\link[=cbind]{cbind()}}
was specified in formula
}
