% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_cor_args}
\alias{check_cor_args}
\title{Checks default arguments for autocorrelation testing}
\usage{
check_cor_args(cor_args)
}
\arguments{
\item{cor_args}{Further arguments for \code{\link[stats:cor]{stats::cor()}}.}
}
\value{
An updated list of \code{\link[stats:cor]{stats::cor()}} arguments
}
\description{
Checks whether all base arguments are present in user-provided
argument list for autocorrelation testing
}
