% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_correlation_threshold}
\alias{check_correlation_threshold}
\title{Check that threshold is between 0-1}
\usage{
check_correlation_threshold(threshold)
}
\arguments{
\item{threshold}{Threshold value to check}
}
\description{
Function to check whether threshold is within appropriate range of >0,<=1
}
