% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{check_for_main_effects}
\alias{check_for_main_effects}
\title{Check for presence of main effects in term}
\usage{
check_for_main_effects(term)
}
\arguments{
\item{term}{The formula to be used with the model.
Can be either quote() or formula().}
}
\value{
Updated term with all main effects
}
\description{
Function that updates a given term to include all main effects
if interactions are present
}
