% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_formula}
\alias{check_formula}
\title{Formula formatting and integrity check}
\usage{
check_formula(formula, data, add_main = TRUE)
}
\arguments{
\item{formula}{The formula to check}

\item{data}{The underlying data for model creation}

\item{add_main}{Whether to add main effect. Default: TRUE}
}
\value{
The updated formula with added main effects if add_main == TRUE
}
\description{
Formats formula for downstream process and checks for main effects
}
