% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_gam_anova}
\alias{check_gam_anova}
\title{Checks if gam formula and evaluation metric are fine}
\usage{
check_gam_anova(formula, evaluation_methods)
}
\arguments{
\item{formula}{Upper formula to be used for model creation/selection}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".}
}
\description{
Checks for gam formulas whether smooth terms are included and
whether user has specified ANOVA as an evaluation metric. If so,
a warning is issued as smooth terms rely on different test statistics than
parametric terms.
}
