% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_plot_interactions}
\alias{check_plot_interactions}
\title{Check whether interactions can be plotted}
\usage{
check_plot_interactions(interactions)
}
\arguments{
\item{interactions}{The interactions to check}
}
\value{
A filtered interactions dataframe supported for plotting
}
\description{
Checks whether given interactions are supported for plotting. Not supported
are interactions of more than two variables and interactions between
factors.
}
