% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_user_args}
\alias{check_user_args}
\title{Checks partial argument matching}
\usage{
check_user_args(call, func)
}
\arguments{
\item{call}{User function call}

\item{func}{Function called by user}
}
\description{
Checks user arguments against list of accepted arguments and throws error
on partial argument matching
}
