% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create model}
\usage{
create_model(
  formula,
  data,
  model_type,
  family = stats::gaussian,
  model_args = list(),
  fit = TRUE
)
}
\arguments{
\item{formula}{The underlying formula for model creation as a language object}

\item{data}{The underlying data for model creation}

\item{model_type}{The type of model to create. Can be either "glm", "lm", "glmer", "lmer",
"gam", "nlme", or "nls"}

\item{family}{The model family (distribution)}

\item{model_args}{Additional arguments that should be given to the model call directly}

\item{fit}{Whether to actually fit the model. Default: TRUE}
}
\value{
A regression model
}
\description{
Creates a regression model
}
\examples{
 # setup
 data("plants")
 # generate an example glm model for plotting
 final_model <- create_model(
  sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    habitat +
    ploidy +
    solar_radiation:isothermality,
  data = plants,
  model_type = "glm",
  family = "quasibinomial"
 )
 @export
}
