% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOptimizer.R
\docType{data}
\name{dataset_info}
\alias{dataset_info}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Lara Kösters \email{lkoesters@bgc-jena.mpg.de}
}
\keyword{data}
