% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_model_family}
\alias{determine_model_family}
\title{Determine appropriate model family}
\usage{
determine_model_family(model_type, data, lhs)
}
\arguments{
\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "lmer", "glmer",
"nlme", "gam", and "nls".}

\item{data}{Data to consider when matching model family}

\item{lhs}{Left-hand-side of model formula (i.e., response term)}
}
\value{
A character vector with valid families given the data and
a notice explaining the choice
}
\description{
Determine appropriate model family
}
