% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_autocorrelations.R
\name{determine_removable_predictors}
\alias{determine_removable_predictors}
\title{Determine predictors to remove}
\usage{
determine_removable_predictors(
  autocorrelations,
  coefficients,
  term_map,
  interactions,
  problematic_predictors = c()
)
}
\arguments{
\item{autocorrelations}{A dataframe with indexed variables and a dataframe with pairs of
autocorrelated variables.}

\item{coefficients}{A dataframe with indexed variables.}

\item{term_map}{Columns to check for autocorrelations. The order of columns dictates
priority basis for removal of predictors. Columns further down the list
are removed first.}

\item{interactions}{Map of columns to terms with only interactions.}

\item{problematic_predictors}{A character vector with columns so far removed due to autocorrelations.}
}
\value{
A list with
a) a vector with removed predictors (NULL if none were removed), and
b) a dataframe with comprehensive information on autocorrelations
}
\description{
Decides which autocorrelated predictors should be removed while
considering interactions and main effects
}
