% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{expand_model_summary}
\alias{expand_model_summary}
\title{A minor version of expanded model summary}
\usage{
expand_model_summary(model_summary, term, categorical_vars, df, round_p = 3)
}
\arguments{
\item{model_summary}{Initial model summary.}

\item{term}{The formula to be used with the model.
Can be either quote() or formula().}

\item{categorical_vars}{List of categorical variables within the model
(base names, i.e., names of columns).}

\item{df}{Dataframe with response and predictors as columns.}

\item{round_p}{Convenience parameter for automatic rounding of p-values.
Default: 3.}
}
\value{
Expanded summary of a regression model with added information
on base variables, variable types and significance levels.
}
\description{
Expand model summary by adding column names and
information on interaction groups.
}
\examples{
# setup
data("plants")

# generate an example glm model
final_model <- glm(
  sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    habitat +
    ploidy +
    solar_radiation:isothermality,
  family = "quasibinomial",
  data = plants
)

# generate an extended model summary with information on predictor columns
# and variable types
models_overview <- stats::coef(summary(final_model))
expand_model_summary(
  models_overview,
  quote(sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    I(isothermality^2) +
    habitat +
    ploidy +
    solar_radiation:annual_mean_temperature +
    solar_radiation:isothermality +
    annual_mean_temperature:isothermality),
  c("habitat", "ploidy"),
  plants,
  3
)
}
