% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{expanded_model_summary}
\alias{expanded_model_summary}
\title{Expansion of model summary}
\usage{
expanded_model_summary(df, model_summary, categorical_vars)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{model_summary}{Summary of model. Either summary(model)$p.table
or stats::coef(summary(model)).}

\item{categorical_vars}{List of categorical variables within the model
(base names, i.e., names of columns)}
}
\value{
Expanded summary of a regression model with added
information on base variables and variable type
}
\description{
Expand model summary by adding column names and
information on interaction groups
}
