% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_categorical_vars}
\alias{extract_categorical_vars}
\title{Extract categorical variables}
\usage{
extract_categorical_vars(model_frame)
}
\arguments{
\item{model_frame}{The \code{\link[=model.frame]{model.frame()}} of a model}
}
\value{
A dataframe of categorical variables with their levels
}
\description{
Extract categorical variables
}
