% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_autocorrelations.R
\name{extract_cor_p_values}
\alias{extract_cor_p_values}
\title{Computes p-values for correlations}
\usage{
extract_cor_p_values(correlations_l, cor_args)
}
\arguments{
\item{correlations_l}{A dataframe with autocorrelations, but pivoted longer}

\item{cor_args}{Further arguments for \code{\link[stats:cor]{stats::cor()}} and \code{\link[stats:cor.test]{stats::cor.test()}}.
Default: method = "pearson" and use = "complete.obs"}
}
\value{
An updated dataframe of autocorrelations with added p-values
}
\description{
Uses \code{\link[stats:cor.test]{stats::cor.test()}} to compute p-values of correlations
detected using \code{\link[stats:cor]{stats::cor()}}.
}
