% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_interactions}
\alias{extract_interactions}
\title{Extract interactions}
\usage{
extract_interactions(predictors)
}
\arguments{
\item{predictors}{A list of predictors extracted from a formula}
}
\value{
A long dataframe with interactions and their separated effects
}
\description{
Extract interactions
}
