% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_terms_per_category}
\alias{extract_terms_per_category}
\title{Extracts numeric and categorical variables as well as interactions related
to a model}
\usage{
extract_terms_per_category(data, model, model_type, family, model_args)
}
\arguments{
\item{data}{The unscaled model data}

\item{model}{The computed model}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "lmer", "glmer", and "gam".}

\item{family}{A character string or call describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Can also be "automatic".}

\item{model_args}{A named list of additional arguments given directly to model call}
}
\value{
A comprehensive dataframe with columns for plotting, numeric and categorical
variables as character vectors as well as
numeric:categorical interactions in a dataframe,
a dataframe containing all traits related to terms relevant to the model.
}
\description{
Extracts numeric and categorical variables as well as interactions related
to a model
}
