% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{find_call}
\alias{find_call}
\title{Split and search model formula}
\usage{
find_call(term, pred_full = NA, return = "all", df_cols = NA)
}
\arguments{
\item{term}{The formula to be used with the model. Can be either quote() or formula()}

\item{pred_full}{Only needed for 'match' and 'negative'. The term to search for within the formula}

\item{return}{Can be 'all', 'atomic', 'interactions', 'main_effects', 'negative', or 'match'. Default: 'all'}

\item{df_cols}{Only needed if return == 'atomic'}
}
\value{
The extracted part(s) of the model formula matching a given pattern. Can be either a call or a vector
}
\description{
Extract part of model formula
}
