% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_numeric_cols}
\alias{formula_numeric_cols}
\title{Extract numeric columns mentioned in formula}
\usage{
formula_numeric_cols(formula, data)
}
\arguments{
\item{formula}{A model formula}

\item{data}{Data with columns mentioned in formula}
}
\value{
Returns all numeric data columns mentioned in formula
}
\description{
Extract numeric columns mentioned in formula
}
