% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_related_cols}
\alias{formula_related_cols}
\title{Extract columns mentioned in formula}
\usage{
formula_related_cols(formula, data)
}
\arguments{
\item{formula}{A model formula}

\item{data}{Data with columns mentioned in formula}
}
\value{
Returns all data columns mentioned in formula
}
\description{
Extract columns mentioned in formula
}
