% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{forward_selection}
\alias{forward_selection}
\title{Forward model selection}
\usage{
forward_selection(
  df,
  model_type,
  term,
  evaluation_methods,
  categorical_vars,
  ...,
  model_family = "binomial",
  trace = FALSE,
  omit.na = "overall"
)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{model_type}{The model to be used (options: (g/n)(l/a)m(er), default: glm)}

\item{term}{The formula to be used with the model. Can be either quote() or formula()}

\item{evaluation_methods}{Methods to be used for model evaluation (options: anova, aic, aicc, bic, default: anova)}

\item{categorical_vars}{List of categorical variables within the model (base names, i.e., names of columns)}

\item{...}{Parameters to be directly used with model call}

\item{model_family}{The family used for glm calculation (default: gaussian)}

\item{trace}{Store and return model selection history (default: FALSE)}

\item{omit.na}{Either 'overall' or 'stepwise'. If 'overall', NAs are removed before modeling. If 'stepwise', NAs are removed per step based on the variables in the current formula}
}
\value{
List containing the final regression model, the significant and marginally significant model variables and the selection history if trace is TRUE
}
\description{
Forward model selection
}
