% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{generate_regression_model}
\alias{generate_regression_model}
\title{Generate regression model}
\usage{
generate_regression_model(df, model_type, term, model_family, ...)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{model_type}{Model type to be used.
Options: "lm", "glm", "lmer", "glmer",
"nlmer", and "gam".}

\item{term}{The formula to be used with the model.
Can be either quote() or formula()}

\item{model_family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options.}

\item{...}{Arguments given to model call}
}
\value{
List with regression model and used model arguments
}
\description{
Helper method to generate a regression model
based on a given model type, family and formula
}
