% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_addable_terms}
\alias{get_addable_terms}
\title{Get addable terms}
\usage{
get_addable_terms(formula, model, model_type)
}
\arguments{
\item{formula}{The full formula to get terms to add from}

\item{model}{A model to determine removable terms for}

\item{model_type}{The type of model to get removable terms for. Can be either "glm", "lm",
"glmer", "lmer", or "gam"}
}
\value{
A dataframe with terms to add and the name of the p-value column
}
\description{
Get addable terms
}
