% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{get_corrected_p_values}
\alias{get_corrected_p_values}
\title{Get corrected p-values}
\usage{
get_corrected_p_values(final_model, postsel, model_type, stat_type = 3)
}
\arguments{
\item{final_model}{The final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{postsel}{The model selected by \code{\link[selcorr:selcorr]{selcorr::selcorr()}}}

\item{model_type}{Model type to be used as character string.
Options: "lm" or "glm"}

\item{stat_type}{Type of Anova test. Default: 3}
}
\value{
A dataframe with model coefficients, and raw and corrected p-values for
filtering of coefficients within the formula
}
\description{
Get corrected p-values
}
