% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_family_character}
\alias{get_family_character}
\title{Family to character}
\usage{
get_family_character(family)
}
\arguments{
\item{family}{The family to turn into a string}
}
\description{
Family to character
}
