% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{get_min_p}
\alias{get_min_p}
\title{Get minimum p-value}
\usage{
get_min_p(coef, coef_type, expanded_model_summary)
}
\arguments{
\item{coef}{Name of coefficient}

\item{coef_type}{Type of coefficient. Can be either 'numeric' or 'categorical'}

\item{expanded_model_summary}{Expanded model summary as returned by \code{\link[=expand_model_summary]{expand_model_summary()}}}
}
\value{
Minimum p-value of provided coefficient
}
\description{
Determine minum p-value of numeric/categorical coefficients
}
