% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_matrix}
\alias{get_model_matrix}
\title{Format data for detection of autocorrelations}
\usage{
get_model_matrix(data, formula, model_type, family, model_args = list())
}
\arguments{
\item{data}{Underlying data for autocorrelation detection and downstream
model creation}

\item{formula}{A formula used for downstream model creation and simplification}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "lmer", "glmer",
"nlme", "gam", and "nls"}

\item{family}{A character string or call describing the family used for model calculation.
See \link[stats:family]{stats::family} for options.}

\item{model_args}{A named list of additional arguments given directly to model call}
}
\value{
Dataframe with interactions, transforms,
and factor variables as numeric columns
}
\description{
Format input data for detection of autocorrelations
by calculating the model matrix that allows autocorrelation testing
for interactions, transforms, and factor variables.
}
