% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_p_values}
\alias{get_model_p_values}
\title{Get p-values of model coefficients}
\usage{
get_model_p_values(model, model_type, type = 3)
}
\arguments{
\item{model}{A model to get p-values for}

\item{model_type}{The type of model to get p-values for. Can be either "glm", "lm", "glmer",
"lmer", "gam"}

\item{type}{Type of Anova}
}
\value{
Dataframe with p-values of predictors
}
\description{
Get p-values of model coefficients
}
