% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stats_test}
\alias{get_stats_test}
\title{Get appropriate statistical test}
\usage{
get_stats_test(model_type, family = stats::gaussian)
}
\arguments{
\item{model_type}{The type of model to get removable terms for. Can be either "glm", "lm",
"glmer", "lmer", or "gam"}

\item{family}{The model family. Default: gaussian}
}
\value{
The appropriate statistical test to be used with add1/drop1
}
\description{
Get appropriate statistical test
}
