% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_term_factors}
\alias{get_term_factors}
\title{Get all factor levels as terms}
\usage{
get_term_factors(model, model_type, model_overview)
}
\arguments{
\item{model}{A model to extract terms from}

\item{model_type}{Model type to be used as character string.
Options: "lm" or "glm"}

\item{model_overview}{Overview of model coefficients with estimates and p-values}
}
\value{
A dataframe with interactions covering at least one categorical variable
with their factor levels and a dataframe with non-interaction
categorical variables with their factor levels
}
\description{
Lists all factor level terms, including in interactions
}
