% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_term_reference_levels}
\alias{get_term_reference_levels}
\title{Omit NAs from model data}
\usage{
get_term_reference_levels(
  categorical_vars,
  data,
  term_map,
  interaction_to_cat,
  contrasts_args = NULL
)
}
\arguments{
\item{categorical_vars}{Character vector with categorical column names relevant to the model}

\item{data}{Underlying model data}

\item{term_map}{A dataframe with column names and corresponding formula terms}

\item{interaction_to_cat}{Dataframe with interactions and their categorical components}

\item{contrasts_args}{Part of model_args that specifies model contrast behaviour}
}
\value{
Dataframe with reference per categorical variables
}
\description{
Omit NAs from model data
}
