% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_interaction_combinations}
\alias{make_interaction_combinations}
\title{Make level combinations of factors within interaction}
\usage{
make_interaction_combinations(x, y)
}
\arguments{
\item{x}{A dataframe group}

\item{y}{The grouping variable(s)}
}
\value{
All combinations of factor levels belonging to an interaction
}
\description{
Make level combinations of factors within interaction
}
