% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_col_to_term}
\alias{map_col_to_term}
\title{Maps model.matrix column names to terms}
\usage{
map_col_to_term(m_matrix, formula)
}
\arguments{
\item{m_matrix}{A model.matrix result}

\item{formula}{A formula used for downstream model creation and simplification}
}
\value{
A dataframe with column names and corresponding formula terms
}
\description{
Maps model.matrix column names to terms
}
