% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{mo_step}
\alias{mo_step}
\title{Take step in model simplification/selection}
\usage{
mo_step(
  df,
  regression_model,
  vars,
  direction,
  evaluation_methods,
  categorical_vars,
  model_type,
  blank_start = FALSE,
  stepwise_omit = FALSE,
  model_args = NA
)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{regression_model}{The current regression model}

\item{vars}{List of model variables that are yet to be added/removed}

\item{direction}{Mode of stepwise model improvement. Either 'forward' (i.e., forward selection), or 'backward' (i.e., backward simplification)}

\item{evaluation_methods}{Methods to be used for model evaluation (options: anova, aic, aicc, bic, default: anova)}

\item{categorical_vars}{List of categorical variables within the model (base names, i.e., names of columns)}

\item{model_type}{The model to be used (options: (g/n)(l/a)m(er), default: glm)}

\item{blank_start}{Used for first run in forward model selection. Creates empty model}

\item{stepwise_omit}{If TRUE, NAs are removed per step based on the variables in the current formula}

\item{model_args}{Arguments given directly to model call}
}
\value{
List containing a) information on whether model was simplified/expanded, b) the (new) current regression model, c) updated list of variables to be added/removed, d) added/removed variable, and e) summary of the model
}
\description{
Used to either remove or add a variable to an already existing regression model
}
