% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{model_improved}
\alias{model_improved}
\title{Regression model was improved}
\usage{
model_improved(
  regression_model,
  old_regression_model,
  evaluation_methods,
  direction,
  model_type
)
}
\arguments{
\item{regression_model}{Regression model 1}

\item{old_regression_model}{Regression model 2}

\item{evaluation_methods}{List of methods to use for model assessment. Options are: anova, aic, aicc, and bic}

\item{direction}{Mode of stepwise model improvement. Either 'forward' (i.e., forward selection), or 'backward' (i.e., backward simplification)}

\item{model_type}{The model to be used (options: (g/n)(l/a)m(er), default: glm)}
}
\value{
Boolean; TRUE if model 1 better fits the data than model 2
}
\description{
Determine if model 1 better fits the data than model 2 given two models and evaluation methods (anova/aic/aicc/bic)
}
