% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_model_is_better}
\alias{new_model_is_better}
\title{New model is better than old}
\usage{
new_model_is_better(assess1, assess2, direction)
}
\arguments{
\item{assess1}{The evaluation results of model 1}

\item{assess2}{The evaluation results of model 2}

\item{direction}{Direction of model selection}
}
\value{
Boolean that describes whether model 1 is better than model 2
}
\description{
New model is better than old
}
