% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{not_enough_valid_cols}
\alias{not_enough_valid_cols}
\title{Stops run when not enough valid columns were detected}
\usage{
not_enough_valid_cols(cols)
}
\arguments{
\item{cols}{Character vector with valid columns for autocorrelation detection}
}
\description{
Stops run when not enough valid columns were detected
}
