% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_categorical_vars}
\alias{plot_categorical_vars}
\title{Plot categorical variables}
\usage{
plot_categorical_vars(
  plot_data,
  categorical_vars,
  response,
  model_overview,
  test = "wilcox",
  plot_type = "boxplot",
  p_threshold = 0.05
)
}
\arguments{
\item{plot_data}{Unscaled data with relevant columns for plotting}

\item{categorical_vars}{A dataframe with categorical variables and corresponding factor levels}

\item{response}{A string representing the response variable}

\item{model_overview}{An overview of model coefficients with estimates and p-values}

\item{test}{Either "t.test" or "wilcox".
Used to calculate statistics for regression plots of categorical variables.
Default: "wilcox"}

\item{plot_type}{Either "boxplot" or "violin".
Used to plot regression plots for categorical variables.
Default: "boxplot"}

\item{p_threshold}{p-value threshold for significance evaluation.
Default: 0.05}
}
\value{
Plots of categorical variables alongside results of statistical tests
}
\description{
Plot categorical variables
}
