% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_effect_size}
\alias{plot_effect_size}
\title{Plot regression model effect sizes}
\usage{
plot_effect_size(models_overview)
}
\arguments{
\item{models_overview}{Custom model summary (see function expanded_model_summary)}
}
\value{
A plot of the effect sizes of the variables listed within the model summary
}
\description{
Plot effect sizes of model variables
}
