% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_estimate}
\alias{plot_estimate}
\title{Plot regression model estimate}
\usage{
plot_estimate(models_overview, regression_model, model_type)
}
\arguments{
\item{models_overview}{Custom model summary (see expanded_model_summary())}

\item{regression_model}{The (simplified) regression model}

\item{model_type}{Type of regression model (e.g., 'glm')}
}
\value{
A plot of the estimates listed within the model summary
}
\description{
Plot estimate, regression and effect size
}
