% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot estimates}
\usage{
plot_estimates(model_overview)
}
\arguments{
\item{model_overview}{Overview of model coefficients with estimates}
}
\value{
An estimate plot
}
\description{
Plot estimates
}
