% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOptimizer.R
\name{plot_glm_vars}
\alias{plot_glm_vars}
\title{GLM response~predictor plotting}
\usage{
plot_glm_vars(
  df,
  response,
  predictors,
  glm_family = "binomial",
  glm_iter = 1000
)
}
\arguments{
\item{df}{A (wide) dataframe with predictors and response as columns}

\item{response}{The response columns within the df}

\item{predictors}{A list of predictors (i.e. column names)}

\item{glm_family}{The family used for glm calculation}

\item{glm_iter}{The number of max iterations used for the glm calculation}
}
\value{
Nothing. Plots relationships in right bottom panel
}
\description{
Plots response~predictor relationship
}
\examples{
data("dataset_info");
coefficients <- c('val_min_gene_length', 'val_max_gene_length',
                  'train_min_gene_length', 'train_max_gene_length',
                  'train_num_samples', 'val_num_samples',
                  'mean_gene_length', 'val_mean_gene_length',
                  'train_mean_gene_length');
plot_glm_vars(dataset_info, "species_confusion", coefficients);
}
