% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_model_features}
\alias{plot_model_features}
\title{Plot model features}
\usage{
plot_model_features(
  regression_model,
  models_overview,
  model_type,
  model_family,
  plot_type = "boxplot",
  jitter_plots = FALSE,
  test = "wilcox",
  round_p = 5,
  remove_insignificant = FALSE
)
}
\arguments{
\item{regression_model}{The (simplified) regression model.}

\item{models_overview}{Custom model summary (see \code{\link[=expand_model_summary]{expand_model_summary()}}).}

\item{model_type}{Model type to be used.
Options: "lm", "glm", "lmer", "glmer",
"nlmer", and "gam".}

\item{model_family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options.}

\item{plot_type}{Whether to plot as boxplot or violin plot
for categorical variables. Default: "boxplot".}

\item{jitter_plots}{Whether to use jitter when generating geom_point plots.
Default: FALSE.}

\item{test}{Test to be used for pairwise significance testing for
categorical variables. Either wilcox or t.test. Default: wilcox.}

\item{round_p}{Convenience parameter for automatic rounding of p-values.
Default: 5.}

\item{remove_insignificant}{Used to exclude insignificant relationships from both estimate
and effect size plots and omit curves from regression plots.
Default: FALSE.}
}
\value{
A list of plots (estimate, regression, and effect size)
and statistics for categorical variables.
}
\description{
Plot estimate, regression and effect size
}
\examples{
# setup
library(dplyr)
data("plants")
model_type <- "glm"
model_family <- "quasibinomial"

# generate an example glm model for plotting
final_model <- glm(
  sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    habitat +
    ploidy +
    solar_radiation:isothermality,
  family = model_family,
  data = plants
)

# generate extended model summary with information on predictor columns
# and variable types
models_overview <- stats::coef(summary(final_model))
models_overview <- LazyModeler::expand_model_summary(
  models_overview,
  quote(sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    I(isothermality^2) +
    habitat +
    ploidy +
    solar_radiation:annual_mean_temperature +
    solar_radiation:isothermality +
    annual_mean_temperature:isothermality),
  c("habitat", "ploidy"),
  plants,
  3
)
models_overview <- models_overview \%>\%
  mutate(effect_direction = case_when(
    Estimate < 0 ~ "negative",
    TRUE ~ "positive"
  ))

# plot the model's features including plotting of estimates,
# numeric and categorical variables using boxplots where possible
plot_model_features(
  final_model,
  models_overview,
  model_type,
  model_family,
  plot_type = "boxplot",
  jitter_plots = TRUE,
  test = "wilcox",
  round_p = 3,
  remove_insignificant = FALSE
)
}
