% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_psi}
\alias{plot_psi}
\title{Plot post model selection inference result}
\usage{
plot_psi(psi_info, p_threshold, label_size = 2.5)
}
\arguments{
\item{psi_info}{Dataframe with raw and corrected p-values of coefficients}

\item{p_threshold}{P-value threshold for significance evaluation}

\item{label_size}{Size of labels within post-selection inference plot.
Default: 2.5}
}
\value{
Plot with raw and corrected p-values of coefficients
}
\description{
Plot post model selection inference result
}
