% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_regression_categorical}
\alias{plot_regression_categorical}
\title{Plot model regression (categorical)}
\usage{
plot_regression_categorical(
  cat_var,
  df,
  models_overview,
  test,
  response_str,
  response_col,
  plot_type,
  regression_plots,
  stat_results
)
}
\arguments{
\item{cat_var}{Categorical variable to plot}

\item{df}{Dataframe object extracted from model}

\item{models_overview}{Custom model summary (see [slightly_expanded_model_summary()])}

\item{test}{Test to be used for pairwise significance testing for categorical variables. Either wilcox or t.test}

\item{response_str}{Response as string}

\item{response_col}{Response column name}

\item{plot_type}{Whether to plot as boxplot or violin plot for categorical variables}

\item{regression_plots}{A list storing regression plots}

\item{stat_results}{A list storing results of statistics}
}
\value{
Regression plots of variables listed within the model summary + statistics
}
\description{
Plot categorical regression of model variables
}
