% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_regression_numeric}
\alias{plot_regression_numeric}
\title{Plot model regression (numeric & mixed)}
\usage{
plot_regression_numeric(
  predictor_full,
  regression_model,
  df,
  term,
  models_overview,
  response_str,
  regression_plots,
  jitter,
  plot_curve,
  round_p,
  model_family,
  model_type
)
}
\arguments{
\item{predictor_full}{Numeric/mixed variable to plot}

\item{regression_model}{The (simplified) regression model}

\item{df}{Dataframe object extracted from model}

\item{term}{The formula to be used with the model}

\item{models_overview}{Custom model summary (see \code{\link[=slightly_expanded_model_summary]{slightly_expanded_model_summary()}})}

\item{response_str}{Response as string}

\item{regression_plots}{A list storing regression plots}

\item{jitter}{Whether to use jitter when generating dotplots}

\item{plot_curve}{Whether to plot regression plot curve using geom_smooth}

\item{round_p}{Convenience parameter for automatic rounding of p-values}

\item{model_family}{Model family used for glm calculation}

\item{model_type}{Type of regression model (e.g., 'glm')}
}
\value{
Regression plots of variables listed within the model summary
}
\description{
Plot numeric/mixed regression of model variables
}
