% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{prepare_num_model}
\alias{prepare_num_model}
\title{Prepare numeric model}
\usage{
prepare_num_model(
  selected_model,
  final_formula,
  final_data,
  model_type,
  family = stats::gaussian,
  model_args = list()
)
}
\arguments{
\item{selected_model}{The model computed during the selection process}

\item{final_formula}{Formula of the final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{final_data}{Data of the final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{model_type}{Model type to be used as character string.
Options: "lm" or "glm"}

\item{family}{A character string or call describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Default: gaussian.}

\item{model_args}{A named list of additional arguments given directly to model call}
}
\value{
A numeric-only version of the final model, where categorical variables
are split into separate columns
}
\description{
Prepare numeric model
}
