% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare data for plotting}
\usage{
prepare_plot_data(model, model_type, model_overview)
}
\arguments{
\item{model}{A model for which to plot features}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "gam"}

\item{model_overview}{An overview of model coefficients with estimates and p-values}
}
\value{
Categorical variables, interactions, and numeric variables
}
\description{
Prepare data for plotting
}
