% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prepare_stats_for_letters}
\alias{prepare_stats_for_letters}
\title{Format statistical results}
\usage{
prepare_stats_for_letters(stat_result, df_sorted, predictor)
}
\arguments{
\item{stat_result}{Initial result of statistical test (wilcox or t.test)}

\item{df_sorted}{Dataframe with sorted mean responses per factor}

\item{predictor}{Predicting factor as string}
}
\value{
Formatted statistical result for letter addition
}
\description{
Formats results of statistical test for addition of factor level letters
}
