% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_autocorrelations.R
\name{remove_autocor_predictors}
\alias{remove_autocor_predictors}
\title{Remove autocorrelated predictors from formula}
\usage{
remove_autocor_predictors(formula, predictors)
}
\arguments{
\item{formula}{Formula to remove predictors from}

\item{predictors}{Predictors to remove}
}
\value{
Updated formula without autocorrelated predictors
}
\description{
Remove autocorrelated predictors from formula
}
