% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_autocorrelations.R
\name{remove_autocorrelations}
\alias{remove_autocorrelations}
\title{Handle autocorrelations}
\usage{
remove_autocorrelations(correlations_w_p, term_map_all, term_map_cor)
}
\arguments{
\item{correlations_w_p}{Dataframe of sorted and indexed autocorrelated variables with p-values.}

\item{term_map_all}{Terms with columns represented in the data.}

\item{term_map_cor}{Terms with columns to check for autocorrelations. The order of columns
dictates priority basis for removal of predictors. Columns further down
the list are removed first.}
}
\value{
A list with
a) a vector with removed predictors (NULL if none were removed), and
b) a dataframe with comprehensive information on autocorrelations
}
\description{
Removes autocorrelations between columns of an input dataframe.
}
