% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removed_preds_to_terms}
\alias{removed_preds_to_terms}
\title{Translate autocorrelated or zero variance data columns to formula terms}
\usage{
removed_preds_to_terms(problematic_predictors, term_map_cor)
}
\arguments{
\item{problematic_predictors}{A vector with data columns to be removed from formula}

\item{term_map_cor}{Terms with columns to check for autocorrelations. The order of columns
dictates priority basis for removal of predictors. Columns further down
the list are removed first.}
}
\value{
A character vector of removable formula predictors
}
\description{
Translate autocorrelated or zero variance data columns to formula terms
}
