% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{run_global_test}
\alias{run_global_test}
\title{Run global statistical test}
\usage{
run_global_test(test, m_frame, response, predictor)
}
\arguments{
\item{test}{Statistical test to run (either kruskal or anova)}

\item{m_frame}{Data for statistical test}

\item{response}{Response variable as string}

\item{predictor}{Predicting factor as string}
}
\value{
Returns boolean for (non-)significant global result
}
\description{
Runs either Kruskal-Wallis or Welch-Anova
}
