% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{run_posthoc_test}
\alias{run_posthoc_test}
\title{Run statistical test}
\usage{
run_posthoc_test(test, m_frame, response, predictor)
}
\arguments{
\item{test}{Statistical test to run (either wilcox or t.test)}

\item{m_frame}{Data for statistical test}

\item{response}{Response variable as string}

\item{predictor}{Predicting factor as string}
}
\value{
Result of statistical test, pivoted from wide to long
}
\description{
Runs either wilcox or t.test
}
