% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{run_psi}
\alias{run_psi}
\title{Run Post-Selection Inference}
\usage{
run_psi(
  selected_model,
  data,
  final_p_values,
  model_type,
  family = stats::gaussian,
  model_args = list(),
  boot_repl = 100,
  k = 2,
  round_p = 6,
  stat_type = 3,
  p_threshold = 0.05,
  label_size = 2.5
)
}
\arguments{
\item{selected_model}{The final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{data}{Data of the model}

\item{final_p_values}{P-values of the final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{model_type}{Model type to be used as character string.
Options: "lm" or "glm"}

\item{family}{A character string or call describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Default: gaussian.}

\item{model_args}{A named list of additional arguments given directly to model call}

\item{boot_repl}{A number or list of psi bootstrap replicates.
Default: 100}

\item{k}{The multiple of the number of degrees of freedom used as
penalty in the model selection. The default k = 2 corresponds to the AIC.}

\item{round_p}{Convenience parameter for automatic rounding of p-values. Default: 5}

\item{stat_type}{Type of Anova test. Default: 3}

\item{p_threshold}{P-value threshold for significance evaluation. Default: 0.05}

\item{label_size}{Size of labels within post-selection inference plot.
Default: 2.5}
}
\value{
A list with a model overview with raw and corrected p-values,
a plot for comparison of raw and corrected p-values,
and the updated model
}
\description{
Runs \code{\link[selcorr:selcorr]{selcorr::selcorr()}} and creates plot for comparison of raw and
corrected p-values
}
