% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{setup_categorical_check}
\alias{setup_categorical_check}
\title{Model variables categorical check}
\usage{
setup_categorical_check(df, categorical_vars, col = quote(coefficients))
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{categorical_vars}{List of categorical variables within the model (base names, i.e., names of columns)}

\item{col}{Column to be used for check. Default is 'coefficients'}
}
\value{
Language object with if/else statements to be applied to model summary with information on whether model variables are categorical
}
\description{
Helper method used for expanding the model summary by adding information on whether or not model variables are categorical
}
\examples{
data("plants");

setup_categorical_check(plants, c('habitat', 'ploidy'), col=quote(predictor))
}
