% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_model.R
\name{simplify_model}
\alias{simplify_model}
\title{Simplify model}
\usage{
simplify_model(
  formula,
  data,
  model_type,
  model_args = list(),
  evaluation_methods = c("anova"),
  direction = "backward",
  family = stats::gaussian,
  p_threshold = 0.05,
  trace = TRUE,
  base_formula = NA
)
}
\arguments{
\item{formula}{Upper formula to be used for model creation/selection}

\item{data}{Underlying model data}

\item{model_type}{The type of model to be created. Can be either "glm", "lm", "glmer", "lmer",
"gam", "nlme", or "nls"}

\item{model_args}{Optional additional arguments to be given directly to the model call}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}

\item{direction}{Direction of model selection. Default: "backward"}

\item{family}{The model family. Default: gaussian}

\item{p_threshold}{p-value threshold for significance evaluation. Default: 0.05}

\item{trace}{Whether to return the selection history. Default. TRUE}

\item{base_formula}{The lower formula used for forward model selection. Only required if
direction = "forward", otherwise this is NA}
}
\value{
The selected model with evaluation metrics and (if trace = TRUE)
the selection history
}
\description{
Wrapper function for backward simplification and forward selection of a model
}
\examples{
# setup
data("plants")

# generate a glm model with the provided term and simplify it
# by applying backward simplification
simplified_model_info <- simplify_model(
  sexual_seed_prop ~ solar_radiation +
    annual_mean_temperature +
    isothermality +
    I(isothermality^2) +
    habitat +
    ploidy +
    solar_radiation:annual_mean_temperature +
    solar_radiation:isothermality +
    annual_mean_temperature:isothermality,
  data = plants,
  model_type = "glm",
  evaluation_methods = c("anova"),
  family = "quasibinomial",
  direction = "backward",
  trace = TRUE
)
}
