% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{slightly_expanded_model_summary}
\alias{slightly_expanded_model_summary}
\title{A minor version of expanded model summary}
\usage{
slightly_expanded_model_summary(
  model_summary,
  response_frm,
  term,
  categorical_check,
  round_p,
  df
)
}
\arguments{
\item{model_summary}{Initial model summary}

\item{response_frm}{Response formula}

\item{term}{The formula to be used with the model. Can be either quote() or formula()}

\item{categorical_check}{Output of \code{\link[=setup_categorical_check]{setup_categorical_check()}}}

\item{round_p}{Convenience parameter for automatic rounding of p-values}

\item{df}{Dataframe with response and predictors as columns}
}
\value{
Expanded summary of a regression model with added information on base variables, variable types and significance levels
}
\description{
Expand model summary by adding column names and information on interaction groups
}
\examples{
data("plants")

model_family = 'quasibinomial'
categorical_check = setup_categorical_check(plants, c('habitat', 'ploidy'), col=quote(predictor))
final_model = glm(sexual_seed_prop ~ altitude + solar_radiation + 
annual_mean_temperature + isothermality + habitat + ploidy + 
solar_radiation:isothermality, family=model_family, data=plants)
models_overview = coef(summary(final_model))
slightly_expanded_model_summary(models_overview, quote(sexual_seed_prop), quote(sexual_seed_prop ~ altitude + solar_radiation + annual_mean_temperature + isothermality + I(isothermality^2) + habitat + ploidy + solar_radiation:annual_mean_temperature + solar_radiation:isothermality + annual_mean_temperature:isothermality), categorical_check, 3, plants)
}
