% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{smaller_expanded_model_summary}
\alias{smaller_expanded_model_summary}
\title{A minor version of expanded model summary}
\usage{
smaller_expanded_model_summary(
  model_summary,
  response_frm,
  term,
  categorical_vars,
  col,
  round_p,
  df
)
}
\arguments{
\item{model_summary}{Initial model summary.}

\item{response_frm}{Response formula.}

\item{term}{The formula to be used with the model.
Can be either quote() or formula().}

\item{categorical_vars}{List of categorical variables within the model
(base names, i.e., names of columns)}

\item{col}{Column to be used for categorical check.}

\item{round_p}{Convenience parameter for automatic rounding of p-values.}

\item{df}{Dataframe with response and predictors as columns.}
}
\value{
Expanded summary of a regression model with added information
on base variables, variable types and significance levels.
}
\description{
Expand model summary by adding column names and
information on interaction groups.
}
\examples{
# setup
data("plants")

# generate an example glm model
final_model <- glm(
  sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    habitat +
    ploidy +
    solar_radiation:isothermality,
  family = "quasibinomial",
  data = plants
)

# generate an extended model summary with information on predictor columns
# and variable types
models_overview <- stats::coef(summary(final_model))
smaller_expanded_model_summary(
  models_overview,
  quote(sexual_seed_prop),
  quote(sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    I(isothermality^2) +
    habitat +
    ploidy +
    solar_radiation:annual_mean_temperature +
    solar_radiation:isothermality +
    annual_mean_temperature:isothermality),
  c("habitat", "ploidy"),
  quote(predictor),
  3,
  plants
)
}
