% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_term_map}
\alias{sort_term_map}
\title{Adds interactions and sorts a term-to-column dataframe by
main effects/interactions}
\usage{
sort_term_map(term_map, sort = TRUE)
}
\arguments{
\item{sort}{Whether to sort the term map by interactions/main effects}

\item{m_matrix}{A dataframe with column names and corresponding formula terms}
}
\value{
A (sorted) term map including information on interactions with
main effects listed first
}
\description{
Adds interactions and sorts a term-to-column dataframe by
main effects/interactions
}
