% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{transform_variables}
\alias{transform_variables}
\title{Transform dataframe variable}
\usage{
transform_variables(
  regression_model,
  df,
  term,
  predictor_full,
  predictor_cols,
  cols_are_numeric
)
}
\arguments{
\item{regression_model}{The (simplified) regression model.}

\item{df}{Dataframe object extracted from model.}

\item{term}{The complete model term.}

\item{predictor_full}{Coefficient as stated in term.}

\item{predictor_cols}{Columns in df representing predictor_full.}

\item{cols_are_numeric}{Boolean information on whether predictor_cols are numeric.}
}
\value{
Returns list with a) dataframe with transformed variables,
b) adjusted predictor_full corresponding to transformed
column name, c) a list with variables with transforms, and
d) the name of a generated column in case of a coefficient interaction.
}
\description{
Transform variable name and values within given dataframe
according to full predictor name and term.
}
