% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{update_final_formula}
\alias{update_final_formula}
\title{Update final formula}
\usage{
update_final_formula(final_formula, final_data, corrected_p_values)
}
\arguments{
\item{final_formula}{Formula of the final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{final_data}{Data of the final model selected in \code{\link[=simplify_model]{simplify_model()}}}

\item{corrected_p_values}{A dataframe with model coefficients, and raw and corrected p-values for
filtering of coefficients within the formula}
}
\value{
An updated model formula and the adjusted corrected_p_values dataframe
with factor levels
}
\description{
Update final formula
}
