% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{update_term}
\alias{update_term}
\title{Updates term after check for autocorrelations}
\usage{
update_term(
  df,
  term,
  autocorrelation_cols,
  autocorrelations,
  autocorrelations_and_preds
)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns.}

\item{term}{The formula to be used with the model. Can be either quote() or formula().}

\item{autocorrelation_cols}{Sorted list of columns in dataframe to be considered when
removing autocorrelations. Should be sorted by priority.
Last element gets eliminated first.}

\item{autocorrelations}{A dataframe with information on autocorrelated variables.}

\item{autocorrelations_and_preds}{Named list with a) a vector containing all removed predictors
(empty if none were removed), and
b) a dataframe containing autocorrelations and
information on removed variables.}
}
\value{
Names list with the updated term, a vector with categorical variables,
and a vector with numeric variables.
}
\description{
Updates term to include all main effects, removes autocorrelations, and
determines categorical/numeric variables
}
