% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_reference_factors}
\alias{add_reference_factors}
\title{Add reference factor levels}
\usage{
add_reference_factors(model_overview, categorical_vars, interactions)
}
\arguments{
\item{model_overview}{Overview of model coefficients with estimates and p-values}

\item{categorical_vars}{Dataframe with categorical variables and factor levels}

\item{interactions}{Dataframe with interactions and underlying main effects}
}
\value{
Updated model overview with information on interactions and categorical
coefficients.
}
\description{
Add reference factor levels
}
