% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{assess_basic_model_quality}
\alias{assess_basic_model_quality}
\title{Assess basic model quality}
\usage{
assess_basic_model_quality(
  final_model,
  plot_quality_assessment,
  model_type,
  model_family
)
}
\arguments{
\item{final_model}{Final, optimized model}

\item{plot_quality_assessment}{Module to use for plots for model quality assessment.
Options: "performance", "baseR". Default: "baseR".}

\item{model_type}{Model type to be used.
Options: "lm", "glm", "lmer", "glmer",
"nlme", and "gam". Default: "glm".}

\item{model_family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Default: "gaussian".}
}
\value{
Plotted model
}
\description{
Creates plots for model quality assessment either using
\code{\link[performance:check_model]{performance::check_model()}} or \code{\link[graphics:plot]{graphics::plot()}}.
}
