% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_model_family}
\alias{check_model_family}
\title{Check validity of model family}
\usage{
check_model_family(family = NULL, automatic = TRUE, data = NULL, lhs = NULL)
}
\arguments{
\item{family}{The chosen model family. Can be NULL if automatic == TRUE. Default: NULL}

\item{automatic}{Whether to automatically choose an appropriate family. Default: NULL}

\item{data}{The underlying data to use when determining an appropriate model family.
Default: NULL}

\item{lhs}{The left-hand-side of the formula used for downstream model creation.
Default: NULL}
}
\value{
An appropriate model family for downstream model creation
}
\description{
Checks whether model family is supported and matches response data of
given formula
}
