% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{check_model_family}
\alias{check_model_family}
\title{Check for plausibility of chosen model family}
\usage{
check_model_family(df, response_frm, model_type, model_family)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns.}

\item{response_frm}{Response formula}

\item{model_type}{Model type to be used.
Options: "lm", "glm", "lmer", "glmer",
"nlme", and "gam". Default: "glm".}

\item{model_family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Default: "gaussian".}
}
\value{
Updated model family
}
\description{
Function that checks whether chosen model family matches response variable
}
