% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_response_data_format}
\alias{check_response_data_format}
\title{Check for data type of response}
\usage{
check_response_data_format(response_col, response_data)
}
\arguments{
\item{response_col}{The response column name}

\item{response_data}{The response data points}
}
\description{
Checks whether formatting of response column matches data
}
