% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_variance}
\alias{check_variance}
\title{Check for variance of relevant formula terms}
\usage{
check_variance(data_matrix, term_map)
}
\arguments{
\item{data_matrix}{Underlying data matrix for autocorrelation detection and downstream
model creation. Should include transforms, interactions, etc.}

\item{term_map}{A dataframe with column names and corresponding formula terms}
}
\value{
A filtered term map including only columns with variance
}
\description{
Checks whether columns corresponding to relevant formula terms
display at least two distinct values.
}
