% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_models}
\alias{compare_models}
\title{Compare models}
\usage{
compare_models(
  evaluation_methods,
  models,
  direction,
  old_model_assessments,
  p_threshold = 0.05,
  delta = 2
)
}
\arguments{
\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}

\item{models}{List with models to compare}

\item{direction}{Direction of model selection used for anova interpretation}

\item{old_model_assessments}{The evaluation results of the previous model}

\item{p_threshold}{p-value threshold for significance evaluation. Default: 0.05}

\item{delta}{Minimal value by which the two models must differ for the newer model to
be rejected}
}
\value{
Information on whether or not the model has improved and the results
of the model's evaluation
}
\description{
Compare models
}
