% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_pivot_longer}
\alias{cor_pivot_longer}
\title{Pivots correlations to long format}
\usage{
cor_pivot_longer(correlations, values_to)
}
\arguments{
\item{correlations}{The result of \code{\link[corrplot:cor.mtest]{corrplot::cor.mtest()}} as a dataframe}

\item{values_to}{Column name for values}
}
\value{
A pivoted dataframe with correlations
}
\description{
Pivots correlations to long format
}
