% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_prep_autocor}
\alias{cor_prep_autocor}
\title{Format correlation information for removal of variables}
\usage{
cor_prep_autocor(autocorrelations, coefficients)
}
\arguments{
\item{autocorrelations}{Dataframe with autocorrelated variables with p-values}

\item{coefficients}{List of variables to consider for removal}
}
\value{
A dataframe with indexed variables and a dataframe with pairs of
autocorrelated variables
}
\description{
Format correlation information for removal of variables
}
