% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_sort_and_filter}
\alias{cor_sort_and_filter}
\title{Sort and filter correlations}
\usage{
cor_sort_and_filter(correlations_w_p, threshold)
}
\arguments{
\item{correlations_w_p}{Dataframe with information on correlations between variables with p-values}

\item{threshold}{The threshold at which to consider two variables autocorrelated}
}
\value{
A sorted dataframe of autocorrelations with a correlation coefficient
equal to or larger than threshold
}
\description{
Sort and filter correlations
}
