% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_model.R
\name{determine_best_pred}
\alias{determine_best_pred}
\title{Determine best predictor to add/remove}
\usage{
determine_best_pred(
  predictors_to_adjust,
  current_formula,
  full_formula,
  data,
  model_type,
  last_model_info,
  direction,
  family = stats::gaussian,
  model_args = list(),
  evaluation_methods = c("anova")
)
}
\arguments{
\item{predictors_to_adjust}{Predictors that can be removed or added at the current selection step}

\item{current_formula}{The formula of the previous model selection step}

\item{full_formula}{The full formula containing all terms}

\item{data}{Underlying data for model creation}

\item{model_type}{The type of model to create. Can be either "glm", "lm", "glmer", "lmer",
"gam", "nlme", or "nls"}

\item{last_model_info}{Assessment information of the previous model selection step}

\item{direction}{Direction of model selection. Default: "backward"}

\item{family}{The model family. Default: gaussian}

\item{model_args}{Optional additional arguments to be given directly to the model call}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}
}
\value{
The best predictor to add/remove, alongside the updated model and
information on whether to continue the selection process or revert to
the previous selection step
}
\description{
Determines the best predictor to add/remove given a list of possible
predictors and methods on which to evaluate the model
}
