% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{determine_model_family}
\alias{determine_model_family}
\title{Determine regression model family}
\usage{
determine_model_family(df, response_frm)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{response_frm}{Response formula}
}
\value{
Model family to be used with regression model
}
\description{
Determine model family automatically based on response column in dataframe
}
