% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_plot_data}
\alias{format_plot_data}
\title{Formats data for plotting}
\usage{
format_plot_data(
  model,
  formula,
  data,
  m_matrix,
  term_map,
  model_type,
  family,
  model_args = list()
)
}
\arguments{
\item{model}{The final model after model selection}

\item{formula}{The final model formula}

\item{data}{The unscaled data; original input to \code{\link[=optimize_model]{optimize_model()}}}

\item{m_matrix}{Matrix of model}

\item{term_map}{A dataframe with column names and corresponding formula terms}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "gam".}

\item{family}{A character string or call describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Can also be "automatic".}

\item{model_args}{A named list of additional model arguments}
}
\value{
Formatted data for plotting with info on categorical and numeric variables
and interactions
}
\description{
Formats data for plotting
}
