% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{generate_regression_model}
\alias{generate_regression_model}
\title{Generate regression model}
\usage{
generate_regression_model(df, model_type, term, model_family, ...)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{model_type}{The model to be used (options: (g/n)(l/a)m(er), default: glm)}

\item{term}{The formula to be used with the model. Can be either quote() or formula()}

\item{model_family}{The family used for glm calculation (default: gaussian)}

\item{...}{Arguments given to model call}
}
\value{
List with regression model and used model arguments
}
\description{
Helper method to generate a regression model based on a given model type, family and formula
}
