% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_removable_terms}
\alias{get_removable_terms}
\title{Get removable terms}
\usage{
get_removable_terms(model, model_type)
}
\arguments{
\item{model}{A model to determine removable terms for}

\item{model_type}{The type of model to get removable terms for. Can be either "glm", "lm",
"glmer", "lmer", "gam"}
}
\value{
A dataframe with removable terms and the name of the p-value column
}
\description{
Get removable terms
}
