% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_autocorrelations.R
\name{handle_autocorrelations}
\alias{handle_autocorrelations}
\title{Handle autocorrelations}
\usage{
handle_autocorrelations(
  data,
  cols,
  remove = TRUE,
  threshold = 0.7,
  cor_args = list(method = c("pearson"), use = "complete.obs")
)
}
\arguments{
\item{data}{Dataframe with response and predictors as columns.}

\item{cols}{Columns to check for autocorrelations. The order of columns dictates
priority basis for removal of predictors. Columns further down the list
are removed first.}

\item{remove}{Boolean switch for automatic removal of autocorrelated variables.}

\item{threshold}{The threshold at which two variables are to be considered as autocorrelated.}

\item{cor_args}{Further arguments for \code{\link[stats:cor]{stats::cor()}}.}
}
\value{
Named list with
a) a vector with removed predictors (NULL if none were removed), and
b) a dataframe with comprehensive information on autocorrelations
}
\description{
Function to process autocorrelations between columns in input dataframe
}
