% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_model.R
\name{nls_nlme}
\alias{nls_nlme}
\title{nls/nlme model assessment}
\usage{
nls_nlme(
  formula,
  data,
  model_type,
  model_args = list(),
  evaluation_methods = c("anova")
)
}
\arguments{
\item{formula}{Formula to be used for model creation}

\item{data}{Underlying model data}

\item{model_type}{The type of model to be created. Can be either "nlme", or "nls"}

\item{model_args}{Optional additional arguments to be given directly to the model call}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}
}
\value{
A model with evaluation metrics
}
\description{
nls and nlme models do not undergo LazyModeler's model selection procedure
but are created and assessed before being returned
}
