% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omit_na_from_model_data}
\alias{omit_na_from_model_data}
\title{Omit NAs from model data}
\usage{
omit_na_from_model_data(formula, data, model_type, family, model_args = c())
}
\arguments{
\item{formula}{Upper formula to be used for model creation/selection}

\item{data}{Underlying model data}

\item{model_type}{The type of model to be created. Can be either "glm", "lm", "glmer", "lmer",
"gam", "nlme", or "nls"}

\item{family}{The model family. Default: gaussian}

\item{model_args}{Optional additional arguments to be given directly to the model call}
}
\value{
Model data without NAs
}
\description{
Omit NAs from model data
}
