% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_model.R
\name{optimize_forward}
\alias{optimize_forward}
\title{Forward model selection}
\usage{
optimize_forward(
  formula,
  base_formula,
  data,
  model_type,
  family = stats::gaussian,
  model_args = list(),
  evaluation_methods = c("anova"),
  p_threshold = 0.05,
  trace = TRUE
)
}
\arguments{
\item{formula}{The upper formula containing all terms to be added}

\item{base_formula}{The lower formula used as a starting point}

\item{data}{Underlying data for model creation}

\item{model_type}{The type of model to create. Can be either "glm", "lm", "glmer", "lmer",
"gam", "nlme", or "nls"}

\item{family}{The model family. Default: gaussian}

\item{model_args}{Optional additional arguments to be given directly to the model call}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}

\item{p_threshold}{p-value threshold for significance evaluation. Default: 0.05}

\item{trace}{Whether to return the selection history. Default. TRUE}
}
\value{
The selected model with evaluation metrics and (if trace = TRUE)
the selection history
}
\description{
Selects a model by sequentially adding terms to a lower formula
}
