% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_model.R
\name{optimize_model}
\alias{optimize_model}
\title{Parent function for model optimization}
\usage{
optimize_model(
  formula,
  data,
  model_type,
  ...,
  family = stats::gaussian,
  model_args = list(),
  evaluation_methods = c("anova"),
  directions = c("backward"),
  simplify_model = TRUE,
  scale_predictors = TRUE,
  detect_autocors = FALSE,
  remove_autocors = FALSE,
  trace = TRUE,
  base_formula = NA,
  use_psi = FALSE,
  quality_assessment = "baseR",
  ac_threshold = 0.7,
  ac_columns = c(),
  cor_args = list(method = c("pearson"), use = "complete.obs"),
  p_threshold = 0.05,
  delta = 2,
  psi_boot_repl = 100,
  psi_k = 2,
  round_p = 5,
  stat_type = 2,
  psi_label_size = 2.5,
  plot_point_position = "jitter",
  plot_relationships = TRUE,
  categorical_stat_test = "wilcox",
  plot_type = "boxplot",
  plot_curve = TRUE
)
}
\arguments{
\item{formula}{The formula to be used with the model. Can be either quote() or formula().}

\item{data}{Dataframe with response and predictors as columns. Note that we require
regular sequential rownames. Custom rownames will be lost.}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "lmer", "glmer",
"nlme", "gam", and "nls".}

\item{...}{Must be empty. Unsupported arguments will result in an error.}

\item{family}{A character string or call describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Can also be "automatic".
Default: gaussian}

\item{model_args}{A named list of additional arguments given directly to model call}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}

\item{directions}{Vector of directions of model selection. Default: c("backward")}

\item{simplify_model}{Whether or not to simplify the full model. Default: TRUE}

\item{scale_predictors}{Whether to apply scaling to predictor variables. Default: FALSE}

\item{detect_autocors}{Whether to detect autocorrelated dataframe variables. Note that we
support autocorrelation detection only for model types "lm", "glm",
"lmer", "glmer", and "gam".
Default: TRUE}

\item{remove_autocors}{Whether to remove autocorrelated dataframe variables from formula.
Default: TRUE}

\item{trace}{Whether to return the full model selection history. Default: TRUE}

\item{base_formula}{Lower formula used for forward model selection. Only required if forward
selection is chosen, otherwise NA.}

\item{use_psi}{Whether to apply post model selection inference to correct p-values.
Default: FALSE}

\item{quality_assessment}{The mode of model quality assessment. Either "baseR" or "performance".
Other values do not produce quality assessment plots.
Default: "baseR"}

\item{ac_threshold}{Threshold at which two variables are to be considered autocorrelated.
Default: 0.7}

\item{ac_columns}{Columns to check for autocorrelations. The order of columns dictates
priority basis for removal of predictors. Columns further down the list
are removed first. If empty, columns are determined based on the given
dataframe. Provided columns need to be part of the formula.
For instance, when providing the formula y ~ x1 + x2 and ac_columns
c("x1", "x2", "x3"), x3 will not be considered for autocorrelation
testing.
Note that LazyModeler still distinguishes between interactions
and main effects. Removal of predictors will start with the evaluation of
correlations between main effects and will remove blocking interactions
when removing a main effect.}

\item{cor_args}{Further arguments for \code{\link[stats:cor]{stats::cor()}} and \code{\link[stats:cor.test]{stats::cor.test()}}.
Default: method = "pearson" and use = "complete.obs"}

\item{p_threshold}{p-value threshold for significance evaluation during autocorrelation
testing and model simplification.
Default: 0.05}

\item{delta}{Used as a minimal distance between model performances that needs to be
present for a candidate model to be considered an improvement over the
last computed model within the model selection process. The logic goes
like this: During both backward and forward selection, the bigger model
needs to show substantially better metrics than the smaller model,
otherwise the smaller model is selected. A bigger delta will require an
even more substantial improvement over the smaller model for the bigger
model to be chosen.
Default: 2}

\item{psi_boot_repl}{A number or list of psi bootstrap replicates.
Default: 100}

\item{psi_k}{The multiple of the number of degrees of freedom used as
penalty in the model selection. The default k = 2 corresponds to the AIC.}

\item{round_p}{Convenience parameter for automatic rounding of p-values. Default: 5}

\item{stat_type}{Type of Anova test}

\item{psi_label_size}{Size of labels within post-selection inference plot.
Default: 2.5}

\item{plot_point_position}{Position adjustment for model feature plots. See the position paramter of
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for more information. Default: "jitter"}

\item{plot_relationships}{Whether to plot regression, effect size, and estimates.
Default: TRUE}

\item{categorical_stat_test}{Either "t.test" or "wilcox". Used to calculate statistics for regression
plots of categorical variables.
Default: "wilcox"}

\item{plot_type}{Either "boxplot" or "violin".
Used to plot regression plots for categorical variables.
Default: "boxplot"}

\item{plot_curve}{Whether to plot \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}} in regression plots. Default: TRUE}
}
\value{
List with a) information on autocorrelated variables and b)
final simplified/expanded models with further information and plots
}
\description{
Optimize model by removing autocorrelations and variables
that do not significantly predict response variable.
}
\examples{
# setup
data("plants")

# generate a glm model with the provided term and simplify it
# by applying backward simplification
simplified_model_info <- optimize_model(
  sexual_seed_prop ~ solar_radiation +
    annual_mean_temperature +
    isothermality +
    I(isothermality^2) +
    habitat +
    ploidy +
    solar_radiation:annual_mean_temperature +
    solar_radiation:isothermality +
    annual_mean_temperature:isothermality,
  data = plants,
  model_type = "glm",
  ac_threshold = 0.8,
  ac_columns = c("solar_radiation",
                 "annual_mean_temperature",
                 "isothermality",
                 "altitude",
                 "latitude_gps_n",
                 "longitude_gps_e"),
  cor_args = list(method = c("spearman"),
                  use = "complete.obs"),
  family = "quasibinomial",
  directions = c("backward"),
  scale_predictors = TRUE,
  evaluation_methods = c("anova"),
  quality_assessment = "performance",
  categorical_stat_test = "wilcox",
  plot_type = "violin",
  round_p = 3,
  trace = TRUE
)
}
