% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_model.R
\name{optimize_model}
\alias{optimize_model}
\title{Parent function for model optimization}
\usage{
optimize_model(
  formula,
  data,
  model_type,
  family = gaussian,
  model_args = list(),
  direction = "backward"
)
}
\arguments{
\item{formula}{The formula to be used with the model. Can be either quote() or formula().}

\item{data}{Dataframe with response and predictors as columns.}

\item{model_type}{Model type to be used.
Options: "lm", "glm", "lmer", "glmer",
"nlme", and "gam". Default: "glm".}

\item{family}{A character string describing the family used for model calculation.
See \link[stats:family]{stats::family} for options. Default: "gaussian".}

\item{model_args}{A named list of additional arguments given directly to model call}
}
\value{
List with a) information on autocorrelated variables and b)
final simplified/expanded models with further information
(see function \code{\link[=simplify_model]{simplify_model()}} for further details)
}
\description{
Optimize model by removing autocorrelations and variables
that do not significantly predict response variable.
}
