% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_model.R
\name{optimizer_step}
\alias{optimizer_step}
\title{Optimization step}
\usage{
optimizer_step(
  formula,
  data,
  model_type,
  last_model_info,
  family = stats::gaussian,
  model_args,
  evaluation_methods,
  p_threshold = 0.05,
  anv_p_value = NULL,
  direction = "backward",
  full_formula = NA,
  candidate = NULL,
  delta = 2
)
}
\arguments{
\item{formula}{Formula to be used for model creation/selection}

\item{data}{Underlying data for model creation}

\item{model_type}{The type of model to create. Can be either "glm", "lm", "glmer", "lmer",
"gam", "nlme", or "nls"}

\item{last_model_info}{Assessment information of the previous model selection step}

\item{family}{The model family. Default: gaussian}

\item{model_args}{Optional additional arguments to be given directly to the model call}

\item{evaluation_methods}{Character vector with methods to use for model evaluation.
Allowed evaluation methods: "aic", "aicc", "bic", or "anova".
Default: c("anova")}

\item{p_threshold}{p-value threshold for significance evaluation. Default: 0.05}

\item{anv_p_value}{(Optional) The p-value calculated by an ANOVA test}

\item{direction}{Direction of model selection. Default: "backward"}

\item{full_formula}{Upper formula to be used with forward model selection.
Should be NA if direction = "backward"}

\item{candidate}{The name of the candidate to add from the formula. Only relevant for GAMs.}

\item{delta}{Used as a minimal distance between model performances that needs to be
present for a candidate model to be considered an improvement over the
last computed model within the model selection process.
Default: 2}
}
\value{
A list covering the newly adjustable terms based on the new model,
model assessments using the specified evaluation methods,
the new model, and the info whether to proceed with the selection process
}
\description{
Takes a step in the model selection process; either forward or backward
}
