% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_categorical_vars}
\alias{plot_categorical_vars}
\title{Plot categorical variables}
\usage{
plot_categorical_vars(
  categorical_vars,
  response,
  m_frame,
  model_overview,
  test = "wilcox",
  plot_type = "boxplot"
)
}
\arguments{
\item{categorical_vars}{A dataframe with categorical variables and corresponding factor levels}

\item{response}{A string representing the response variable}

\item{m_frame}{A model frame. See \code{\link[stats:model.frame]{stats::model.frame()}}}

\item{model_overview}{An overview of model coefficients withe estimates and p-values}

\item{test}{Either "t.test" or "wilcox".
Used to calculate statistics for regression plots of categorical variables.
Default: "wilcox"}

\item{plot_type}{Either "boxplot" or "violin".
Used to plot regression plots for categorical variables.
Default: "boxplot"}
}
\value{
Plots of categorical variables alongside results of statistical tests
}
\description{
Plot categorical variables
}
