% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_effect_sizes}
\alias{plot_effect_sizes}
\title{Plot effect sizes}
\usage{
plot_effect_sizes(model_overview, formatted_labels)
}
\arguments{
\item{model_overview}{Overview of model coefficients with estimates}

\item{formatted_labels}{A list with formatted labels (italic and bold for factor levels)}
}
\value{
An effect size plot
}
\description{
Plot effect sizes
}
