% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model}
\alias{plot_model}
\title{Plot model}
\usage{
plot_model(
  model,
  data,
  model_type,
  family,
  model_args = list(),
  quality_assessment = "baseR",
  test = "wilcox",
  plot_type = "boxplot",
  plot_curve = TRUE,
  round_p = 5,
  point_position = "jitter"
)
}
\arguments{
\item{model}{A model for which to plot features}

\item{data}{Unscaled data underlying model}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "gam"}

\item{quality_assessment}{The mode of model quality assessment. Either "baseR" or "performance".
Default: "baseR"}

\item{test}{Either "t.test" or "wilcox".
Used to calculate statistics for regression plots of categorical variables.
Default: "wilcox"}

\item{plot_type}{Either "boxplot" or "violin".
Used to plot regression plots for categorical variables.
Default: "boxplot"}

\item{plot_curve}{Whether to plot \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}} in regression plots. Default: TRUE}

\item{round_p}{Convenience parameter for automatic rounding of p-values. Default: 5}

\item{point_position}{Position adjustment for model feature plots. See the position paramter of
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for more information}
}
\value{
A list of plots (estimate, regression, and effect size)
and statistics for categorical variables.
}
\description{
Plots model features such as estimates, effect sizes, and plots
visualizing the relationship between the response variable and predictors
}
\examples{
 # setup
 data("plants")
 # generate an example glm model for plotting
 final_model <- create_model(
  sexual_seed_prop ~ altitude +
    solar_radiation +
    annual_mean_temperature +
    isothermality +
    habitat +
    ploidy +
    solar_radiation:isothermality,
  data = plants,
  model_type = "glm",
  family = "quasibinomial"
 )
 # plot the model's features including plotting of estimates,
 # numeric and categorical variables using boxplots where possible
 p <- plot_model(
  final_model,
  model_type = "glm",
  plot_type = "boxplot",
  point_position = "jitter",
  test = "wilcox",
  round_p = 3
 )
}
