% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_model_features}
\alias{plot_model_features}
\title{Plot model features}
\usage{
plot_model_features(
  regression_model,
  models_overview,
  model_type,
  model_family,
  plot_type = "boxplot",
  jitter_plots = FALSE,
  test = "wilcox",
  round_p = 5,
  remove_insignificant = FALSE
)
}
\arguments{
\item{regression_model}{The (simplified) regression model}

\item{models_overview}{Custom model summary (see \code{\link[=slightly_expanded_model_summary]{slightly_expanded_model_summary()}})}

\item{model_type}{Type of regression model (e.g., 'glm')}

\item{model_family}{Model family used for glm calculation (default: gaussian)}

\item{plot_type}{Whether to plot as boxplot or violin plot for categorical variables (default: boxplot)}

\item{jitter_plots}{Whether to use jitter when generating geom_point plots (default: FALSE)}

\item{test}{Test to be used for pairwise significance testing for categorical variables. Either wilcox or t.test (default: wilcox)}

\item{round_p}{Convenience parameter for automatic rounding of p-values}

\item{remove_insignificant}{Used to exclude insignificant relationships from both estimate and effect size plots and omit curves from regression plots (default: FALSE)}
}
\value{
A list of plots (estimate, regression, and effect size) and statistics for categorical variables
}
\description{
Plot estimate, regression and effect size
}
\examples{
data("plants");

model_type = 'glm'
model_family = 'quasibinomial'
categorical_check = setup_categorical_check(plants, c('habitat', 'ploidy'), col=quote(predictor))
final_model = glm(sexual_seed_prop ~ altitude + solar_radiation + annual_mean_temperature + isothermality + habitat + ploidy + solar_radiation:isothermality, family=model_family, data=plants)

models_overview = coef(summary(final_model))
models_overview = slightly_expanded_model_summary(models_overview, quote(sexual_seed_prop), quote(sexual_seed_prop ~ altitude + solar_radiation + annual_mean_temperature + isothermality + I(isothermality^2) + habitat + ploidy + solar_radiation:annual_mean_temperature + solar_radiation:isothermality + annual_mean_temperature:isothermality), categorical_check, 3, plants)
models_overview = models_overview |>
dplyr::mutate(effect_direction = dplyr::case_when(
Estimate < 0 ~ 'negative',
TRUE ~ 'positive'
))

plot_model_features(final_model, models_overview, model_type, model_family, plot_type='boxplot', jitter_plots=TRUE, test='wilcox', round_p=3, remove_insignificant=FALSE)
}
