% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_numeric_vars}
\alias{plot_numeric_vars}
\title{Plot numeric variables}
\usage{
plot_numeric_vars(
  numeric_vars,
  response,
  m_frame,
  m_matrix,
  model_overview,
  model_type,
  family,
  plot_curve = TRUE,
  round_p = 5,
  point_position = "jitter",
  group_var = NA
)
}
\arguments{
\item{numeric_vars}{A vector with names of numeric variables}

\item{response}{A string representing the response variable}

\item{m_frame}{A model frame. See \code{\link[stats:model.frame]{stats::model.frame()}}}

\item{m_matrix}{A model frame. See \code{\link[stats:model.matrix]{stats::model.matrix()}}}

\item{model_overview}{An overview of model coefficients withe estimates and p-values}

\item{model_type}{Model type to be used as character string.
Options: "lm", "glm", "gam"}

\item{family}{The model family}

\item{plot_curve}{Whether to plot \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}} in regression plots. Default: TRUE}

\item{round_p}{Convenience parameter for automatic rounding of p-values. Default: 5}

\item{point_position}{Position adjustment for model feature plots. See the position paramter of
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for more information. Default: "jitter"}

\item{group_var}{A grouping variable for coloring}
}
\value{
Plots of numeric variables
}
\description{
Plot numeric variables
}
