% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{plot_psi}
\alias{plot_psi}
\title{Plot PSI p-value changes}
\usage{
plot_psi(psi_df, p_threshold = 0.05, label_size = 2.5)
}
\arguments{
\item{psi_df}{Dataframe provided by \code{\link[=apply_psi]{apply_psi()}}.}

\item{p_threshold}{P-value threshold to be used for significance filtering.}

\item{label_size}{Size of test labels within plot.}
}
\value{
Plot displaying p-values before and after post selection inference
}
\description{
Plot p-values of model coefficients before and
after post selection inference.
}
