% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{remove_autocorrelations}
\alias{remove_autocorrelations}
\title{Autocorrelation identification and removal}
\usage{
remove_autocorrelations(
  df,
  coefficients,
  ...,
  automatic_removal = TRUE,
  autocorrelation_threshold = 0.7,
  correlation_method = "pearson"
)
}
\arguments{
\item{df}{Dataframe with response and predictors as columns}

\item{coefficients}{List of predictors sorted by their relevance (most to least)}

\item{...}{Arguments given directly to 'cor' method}

\item{automatic_removal}{Determines whether to automatically remove autocorrelations}

\item{autocorrelation_threshold}{Threshold at which two variables are considered autocorrelated}

\item{correlation_method}{Method used for correlation calculation}
}
\value{
Named list with a) a vector containing all removed predictors (empty if none were removed), and b) a dataframe containing autocorrelations and information on removed variables
}
\description{
Identify autocorrelated variables. Automatically remove autocorrelated variables if specified
}
