% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOptimizer.R
\name{remove_autocorrelations}
\alias{remove_autocorrelations}
\title{Autocorrelation removal}
\usage{
remove_autocorrelations(
  df,
  coefficients,
  automatic_removal = TRUE,
  autocorrelation_threshold = 0.8,
  correlation_method = "pearson"
)
}
\arguments{
\item{df}{A (wide) dataframe with predictors and response as columns}

\item{coefficients}{A list of predictors sorted by their relevance (most to least)}

\item{automatic_removal}{A boolean to decide whether to automatically remove autocorrelations or leave it to the user to do so manually}

\item{autocorrelation_threshold}{The threshold deciding (together with the p-values) whether two variables are autocorrelated}

\item{correlation_method}{The method used for correlation calculation}
}
\value{
An edited list of coefficients where autocorrelated variables are removed (according to sorted list of coefficients)
}
\description{
Remove autocorrelations
}
\examples{
data("dataset_info");
coefficients <- c('val_min_gene_length', 'val_max_gene_length',
                  'train_min_gene_length', 'train_max_gene_length',
                  'train_num_samples', 'val_num_samples',
                  'mean_gene_length', 'val_mean_gene_length',
                  'train_mean_gene_length');
remove_autocorrelations(df, coefficients, automatic_removal = TRUE);
}
