% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{run_stats}
\alias{run_stats}
\title{Compute statistics}
\usage{
run_stats(m_frame, response, predictor, test = "wilcox")
}
\arguments{
\item{m_frame}{Data for test}

\item{response}{Response variable as string}

\item{predictor}{Predicting factor as string}

\item{test}{Test to run. Either "wilcox" or "t.test"}
}
\value{
List with
a) updated m_frame with letters indicating significant differences
between factor levels
b) result of statistical test
}
\description{
Runs either wilcox or t.test
}
