% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelOptimizer.R
\name{simplify_model}
\alias{simplify_model}
\title{Model simplification}
\usage{
simplify_model(glm_model, aic_bic = FALSE, model_method = "binomial")
}
\arguments{
\item{glm_model}{Starting glm model that will be simplified}

\item{aic_bic}{Whether to apply AIC and BIC difference in addition to ANOVA to determine if model is better/worse than ancestor}
}
\value{
A log with models + statistics as well as final model
}
\description{
Simplify models
}
\examples{
data("dataset_info");
coefficients <- c('val_min_gene_length', 'val_max_gene_length',
                  'train_min_gene_length', 'train_max_gene_length',
                  'train_num_samples', 'val_num_samples',
                  'mean_gene_length', 'val_mean_gene_length',
                  'train_mean_gene_length');
frm <- as.formula(paste("species_confusion", "~", paste(coefficients, collapse='+')))
glm_model <- eval(bquote(glm(as.formula(frm), data=.(dataset_info), family = .(glm_family))))
simplify_model(glm_model, aic_bic=TRUE);
}
