% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyModeler.R
\name{transform_variables}
\alias{transform_variables}
\title{Transform dataframe variable}
\usage{
transform_variables(
  regression_model,
  models_overview,
  df,
  term,
  predictor_full,
  main_effects,
  pred_data_type
)
}
\arguments{
\item{regression_model}{The (simplified) regression model.}

\item{models_overview}{Custom model summary (see \code{\link[=expand_model_summary]{expand_model_summary()}}).}

\item{df}{Dataframe object extracted from model.}

\item{term}{The complete model term.}

\item{predictor_full}{Coefficient as stated in term.}

\item{main_effects}{Main effects included in predictor_full.}

\item{pred_data_type}{Information on whether main_effects are numeric.
Either: "numeric", "categorical", or "mixed".}
}
\value{
Returns list with a) dataframe with transformed variables,
b) adjusted predictor_full corresponding to transformed
column name, c) a list with variables with transforms, and
d) the name of a generated column in case of a coefficient interaction.
}
\description{
Transform variable name and values within given dataframe
according to full predictor name and term.
}
